scripts/tdr/copy_from_tdr_to_gcs_hca/requirements.txt
# venv setup
# python3 -m venv venv
# source venv/bin/activate
# pip install -r requirements.txt

requests
google-cloud
google-auth
